Change signature of enabled_when_unfrozen
This corrects the incorrect signature that enabled_when_unfrozen had.
The new signature is (method: Callable) -> Any.

Previously, enabled_when_unfrozen had the type signature
(method: _F) -> _F where _F is a type variable bounded by Callable.
This signature was chosen because it preserves the signature of the
original function. However, this signature is incorrect because if the
user gives a callable instance of a user type, the function will return
a user function, not an instance of that user type.

In the future, the signature of this function will be made to be more
specific using `ParamSpec` from the third-party `typing-extensions`
package. This is to preserve the signature of the given function.